requirements.txt
annotated-types==0.7.0
anyio==4.5.0
asttokens==2.4.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docopt==0.6.2
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.115.0
fastapi-cli==0.0.5
fastjsonschema==2.20.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
ipython==8.12.3
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
logzero==1.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mongoengine==0.29.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
openpyxl==3.1.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
ply==3.11
prompt_toolkit==3.0.47
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.10.1
pyotp==2.9.0
pyparsing==3.1.4
pysmi==1.5.0
pysnmp==7.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==7.0.0
rdflib-jsonld==0.6.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
setuptools==75.1.0
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smartapi-python==1.5.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.38.5
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
Twisted==24.7.0
txaio==23.1.1
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.1.0
uvicorn==0.30.6
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
yarg==0.1.9
zope.interface==7.0.3
